+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/chatcommands/impl/party/CoordsCommand.kt
@@ -3,8 +3,9 @@ package me.nobaboy.nobaaddons.features.chat.chatcommands.impl.party
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatCommand
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatContext
 import me.nobaboy.nobaaddons.utils.LocationUtils
+import kotlin.time.Duration.Companion.seconds
 
-class CoordsCommand : ChatCommand() {
+class CoordsCommand : ChatCommand(0.3.seconds) {
 	override val enabled: Boolean get() = config.party.coords
 
 	override val name: String = "coords"