+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/filters/miscellaneous/ProfileInfoChatFilter.kt
@@ -1,31 +1,34 @@
 package me.nobaboy.nobaaddons.features.chat.filters.miscellaneous
 
+//? if <1.21.5 {
+/*import net.minecraft.text.ClickEvent*/
+//?}
+
 import me.nobaboy.nobaaddons.api.skyblock.SkyBlockAPI
 import me.nobaboy.nobaaddons.features.chat.filters.IChatFilter
 import me.nobaboy.nobaaddons.repo.Repo
 import me.nobaboy.nobaaddons.repo.Repo.fromRepo
 import me.nobaboy.nobaaddons.utils.CommonPatterns
 import me.nobaboy.nobaaddons.utils.RegexUtils.anyFullMatch
-import net.minecraft.text.ClickEvent
 import net.minecraft.text.Text
 
 object ProfileInfoChatFilter : IChatFilter {
 	private val profileInfoPatterns: List<Regex> by Repo.list(
 		Regex("^You are playing on profile: [A-z-() ]+").fromRepo("filter.profile.playing_on"),
-		Regex("^Profile ID: [A-z0-9-]+").fromRepo("filter.profile.id"),
+		Regex("^Profile ID: [A-z0-9-]+").fromRepo("filter.profile.uuid"),
 	)
 
-	private val suggestPattern by Regex("^§e§lCLICK THIS TO SUGGEST IT IN CHAT §7§l\\[(?:NO )?DASHES]").fromRepo("filter.profile.suggest")
+	private val UUID_SUGGEST_REGEX by Regex("^§e§lCLICK THIS TO SUGGEST IT IN CHAT §7§l\\[(?:NO )?DASHES]").fromRepo("filter.profile.uuid_suggest")
 
 	private fun isSuggestProfile(message: Text): Boolean {
-		if(suggestPattern.matches(message.string)) {
+		if(UUID_SUGGEST_REGEX.matches(message.string)) {
 			val clickAction = message.style.clickEvent ?: return false
 			//? if >=1.21.5-pre2 {
 			/*val value = (clickAction as? ClickEvent.SuggestCommand)?.command ?: return false
 			*///?} else {
 			val value = clickAction.value
 			//?}
-			return CommonPatterns.UUID.matches(value)
+			return CommonPatterns.UUID_REGEX.matches(value)
 		}
 		return false
 	}