+++ src/main/kotlin/me/nobaboy/nobaaddons/features/inventory/slotinfo/items/KuudraKeyTierSlotInfo.kt
@@ -1,10 +1,10 @@
-package me.nobaboy.nobaaddons.features.visuals.slotinfo.items
+package me.nobaboy.nobaaddons.features.inventory.slotinfo.items
 
 import me.nobaboy.nobaaddons.events.impl.render.ScreenRenderEvents
-import me.nobaboy.nobaaddons.features.visuals.slotinfo.ISlotInfo
+import me.nobaboy.nobaaddons.features.inventory.slotinfo.ISlotInfo
 import me.nobaboy.nobaaddons.utils.items.ItemUtils.asSkyBlockItem
 
-object KuudraKeyTierInfoSlot : ISlotInfo {
+object KuudraKeyTierSlotInfo : ISlotInfo {
 	override val enabled: Boolean get() = config.kuudraKeyTier
 
 	override fun handle(event: ScreenRenderEvents.DrawItem) {