+++ src/main/kotlin/me/nobaboy/nobaaddons/features/inventory/slotinfo/items/MinionTierSlotInfo.kt
@@ -1,4 +1,4 @@
-package me.nobaboy.nobaaddons.features.visuals.slotinfo.items
+package me.nobaboy.nobaaddons.features.inventory.slotinfo.items
 
 object MinionTierSlotInfo : AbstractItemIdTierSlotInfo() {
 	override val pattern = Regex("^[A-Z_]+_GENERATOR_(?<tier>\\d+)")