+++ src/main/kotlin/me/nobaboy/nobaaddons/features/inventory/slotinfo/uielements/SkyBlockLevelSlotInfo.kt
@@ -1,8 +1,8 @@
-package me.nobaboy.nobaaddons.features.visuals.slotinfo.uielements
+package me.nobaboy.nobaaddons.features.inventory.slotinfo.uielements
 
 import me.nobaboy.nobaaddons.api.skyblock.SkyBlockAPI
 import me.nobaboy.nobaaddons.events.impl.render.ScreenRenderEvents
-import me.nobaboy.nobaaddons.features.visuals.slotinfo.ISlotInfo
+import me.nobaboy.nobaaddons.features.inventory.slotinfo.ISlotInfo
 import me.nobaboy.nobaaddons.utils.InventoryUtils
 
 object SkyBlockLevelSlotInfo : ISlotInfo {